feat: nav items styling
